Implement PriorityFeesButton Move PriorityFeesButton and RewardsButton into header components file
