+++ src/Layout/components/Header/Header.tsx
@@ -1,19 +1,15 @@
-import { useEffect, useState } from 'react'
-
 import { useWallet } from '@solana/wallet-adapter-react'
-import classNames from 'classnames'
 import { NavLink } from 'react-router-dom'
-import TextTransition, { presets } from 'react-text-transition'
 
 import { BanxNotificationsButton } from '@banx/components/BanxNotifications'
-import { Button, WalletConnectButton } from '@banx/components/Buttons'
+import { WalletConnectButton } from '@banx/components/Buttons'
 
-import { Cup, Logo, LogoFull } from '@banx/icons'
+import { Logo, LogoFull } from '@banx/icons'
 import { PATHS } from '@banx/router'
 
 import { BurgerIcon } from '../BurgerMenu'
-import { isActivePath } from '../Navbar/helpers'
 import ThemeSwitcher from '../ThemeSwitcher'
+import { PriorityFeesButton, RewardsButton } from './components'
 
 import styles from './Header.module.less'
 
@@ -30,42 +26,10 @@ export const Header = () => {
         <RewardsButton />
         <ThemeSwitcher />
         {connected && <BanxNotificationsButton />}
+        <PriorityFeesButton />
         <WalletConnectButton />
       </div>
       <BurgerIcon />
     </div>
-  )
-}
-
-const RewardsButton = () => {
-  const TOKENS = [{ text: '$BANX' }, { text: '$BONK', style: styles.rewardsButtonTextOrange }]
-
-  const [index, setIndex] = useState(0)
-
-  useEffect(() => {
-    const intervalId = setInterval(() => setIndex((index) => index + 1), 2500)
-    return () => clearTimeout(intervalId)
-  }, [])
-
-  const currentToken = TOKENS[index % 2]
-
-  return (
-    <NavLink to={PATHS.LEADERBOARD}>
-      <Button
-        type="circle"
-        variant="text"
-        className={classNames(styles.rewardsButton, {
-          [styles.active]: isActivePath(PATHS.LEADERBOARD),
-        })}
-      >
-        <Cup />
-        <div className={styles.rewardsButtonText}>
-          Farm{' '}
-          <TextTransition springConfig={presets.wobbly} className={currentToken.style}>
-            {currentToken.text}
-          </TextTransition>
-        </div>
-      </Button>
-    </NavLink>
   )
 }
